implementacion de la HU1
