--- src/main/java/BootcampPragma/emazon/EmazonApplication.java
@@ -1,13 +1,0 @@
-package BootcampPragma.emazon;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EmazonApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EmazonApplication.class, args);
-	}
-
-}
