--- src/main/java/BootcampPragma/emazon/aplication/dto/CategoryResponse.java
@@ -1,12 +1,0 @@
-package BootcampPragma.emazon.aplication.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class CategoryResponse {
-    private Long id;
-    private String name;
-    private String description;
-}
