--- src/main/java/BootcampPragma/emazon/aplication/handler/ICategoryHandler.java
@@ -1,9 +1,0 @@
-package BootcampPragma.emazon.aplication.handler;
-
-import BootcampPragma.emazon.aplication.dto.CategoryRequest;
-
-public interface ICategoryHandler {
-
-    void saveCategory(CategoryRequest categoryRequest);
-
-}
