--- src/main/java/BootcampPragma/emazon/domain/api/ICategoryServicePort.java
@@ -1,10 +1,0 @@
-package BootcampPragma.emazon.domain.api;
-
-
-import BootcampPragma.emazon.domain.entity.Category;
-
-public interface ICategoryServicePort {
-
-    void saveCategory(Category category);
-
-}
