--- src/main/java/BootcampPragma/emazon/domain/spi/ICategoryPersistencePort.java
@@ -1,11 +1,0 @@
-package BootcampPragma.emazon.domain.spi;
-
-
-import BootcampPragma.emazon.domain.entity.Category;
-
-public interface ICategoryPersistencePort {
-
-    void saveCategory(Category category);
-
-
-}
