--- src/main/java/BootcampPragma/emazon/infrastructure/exception/CategoryAlreadyExistException.java
@@ -1,7 +1,0 @@
-package BootcampPragma.emazon.infrastructure.exception;
-
-public class CategoryAlreadyExistException extends RuntimeException {
-    public CategoryAlreadyExistException() {
-        super();
-    }
-}
