--- src/main/java/BootcampPragma/emazon/infrastructure/input/rest/CategoryRestController.java
@@ -1,25 +1,0 @@
-package BootcampPragma.emazon.infrastructure.input.rest;
-
-import BootcampPragma.emazon.aplication.dto.CategoryRequest;
-import BootcampPragma.emazon.aplication.handler.ICategoryHandler;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-
-@RestController
-@RequestMapping("/category")
-@RequiredArgsConstructor
-public class CategoryRestController {
-
-    private final ICategoryHandler categoryHandler;
-
-    @PostMapping
-    public ResponseEntity<Void> saveCategory(@Valid @RequestBody CategoryRequest categoryRequest) {
-        categoryHandler.saveCategory(categoryRequest);
-        return ResponseEntity.status(HttpStatus.CREATED).build();
-    }
-
-}
