--- src/main/java/BootcampPragma/emazon/infrastructure/output/jpa/entity/CategoryEntity.java
@@ -1,22 +1,0 @@
-package BootcampPragma.emazon.infrastructure.output.jpa.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Table(name = "category")
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@Setter
-public class CategoryEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String name;
-    private String description;
-
-}
