--- src/main/java/BootcampPragma/emazon/infrastructure/output/repository/ICategoryRepository.java
@@ -1,9 +1,0 @@
-package BootcampPragma.emazon.infrastructure.output.repository;
-
-import BootcampPragma.emazon.infrastructure.output.jpa.entity.CategoryEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ICategoryRepository extends JpaRepository<CategoryEntity, Long> {
-    // Método para verificar si existe una categoría por nombre
-    boolean existsByName(String name);
-}
