+++ src/main/java/bootcampragma/emazon/EmazonApplication.java
@@ -1,0 +1,13 @@
+package bootcampragma.emazon;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class EmazonApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(EmazonApplication.class, args);
+	}
+
+}
