+++ src/main/java/bootcampragma/emazon/aplication/dto/CategoryResponse.java
@@ -1,0 +1,12 @@
+package bootcampragma.emazon.aplication.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class CategoryResponse {
+    private Long id;
+    private String name;
+    private String description;
+}
