+++ src/main/java/bootcampragma/emazon/aplication/handler/CategoryHandler.java
@@ -1,0 +1,24 @@
+package bootcampragma.emazon.aplication.handler;
+
+import bootcampragma.emazon.aplication.dto.CategoryRequest;
+import bootcampragma.emazon.aplication.mapper.CategoryRequestMapper;
+import bootcampragma.emazon.domain.api.ICategoryServicePort;
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+
+@Service
+@RequiredArgsConstructor
+@Transactional
+public class CategoryHandler implements ICategoryHandler {
+
+    private final ICategoryServicePort categoryServicePort;
+    private final CategoryRequestMapper categoryRequestMapper;
+
+    @Override
+    public void saveCategory(CategoryRequest categoryRequest) {
+        categoryServicePort.saveCategory(categoryRequestMapper.toRequest(categoryRequest));
+    }
+
+}
