+++ src/main/java/bootcampragma/emazon/aplication/handler/ICategoryHandler.java
@@ -1,0 +1,9 @@
+package bootcampragma.emazon.aplication.handler;
+
+import bootcampragma.emazon.aplication.dto.CategoryRequest;
+
+public interface ICategoryHandler {
+
+    void saveCategory(CategoryRequest categoryRequest);
+
+}
