+++ src/main/java/bootcampragma/emazon/aplication/mapper/CategoryRequestMapper.java
@@ -1,0 +1,12 @@
+package bootcampragma.emazon.aplication.mapper;
+
+import bootcampragma.emazon.aplication.dto.CategoryRequest;
+import bootcampragma.emazon.domain.entity.Category;
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface CategoryRequestMapper {
+
+    Category toRequest(CategoryRequest categoryRequest);
+
+}
