+++ src/main/java/bootcampragma/emazon/domain/api/ICategoryServicePort.java
@@ -1,0 +1,10 @@
+package bootcampragma.emazon.domain.api;
+
+
+import bootcampragma.emazon.domain.entity.Category;
+
+public interface ICategoryServicePort {
+
+    void saveCategory(Category category);
+
+}
