+++ src/main/java/bootcampragma/emazon/domain/spi/ICategoryPersistencePort.java
@@ -1,0 +1,11 @@
+package bootcampragma.emazon.domain.spi;
+
+
+import bootcampragma.emazon.domain.entity.Category;
+
+public interface ICategoryPersistencePort {
+
+    void saveCategory(Category category);
+
+
+}
