+++ src/main/java/bootcampragma/emazon/infrastructure/configuration/BeanConfiguration.java
@@ -1,0 +1,31 @@
+package bootcampragma.emazon.infrastructure.configuration;
+
+import bootcampragma.emazon.domain.api.ICategoryServicePort;
+import bootcampragma.emazon.domain.spi.ICategoryPersistencePort;
+import bootcampragma.emazon.domain.usecase.CategoryUseCase;
+import bootcampragma.emazon.infrastructure.output.jpa.adapter.CategoryJpaAdapter;
+import bootcampragma.emazon.infrastructure.output.jpa.mapper.CategoryEntityMapper;
+import bootcampragma.emazon.infrastructure.output.repository.ICategoryRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@RequiredArgsConstructor
+
+public class BeanConfiguration {
+
+    private final ICategoryRepository categoryRepository;
+    private final CategoryEntityMapper categoryEntityMapper;
+
+    @Bean
+    public ICategoryPersistencePort categoryPersistencePort() {
+        return new CategoryJpaAdapter(categoryRepository, categoryEntityMapper);
+    }
+
+    @Bean
+    public ICategoryServicePort categoryServicePort(){
+        return new CategoryUseCase(categoryPersistencePort());
+    }
+
+}
