+++ src/main/java/bootcampragma/emazon/infrastructure/exception/CategoryAlreadyExistException.java
@@ -1,0 +1,7 @@
+package bootcampragma.emazon.infrastructure.exception;
+
+public class CategoryAlreadyExistException extends RuntimeException {
+    public CategoryAlreadyExistException() {
+        super("There is already a category with that name");
+    }
+}
