+++ src/main/java/bootcampragma/emazon/infrastructure/exception/CategoryDescriptionEmptyException.java
@@ -1,0 +1,7 @@
+package bootcampragma.emazon.infrastructure.exception;
+
+public class CategoryDescriptionEmptyException extends RuntimeException{
+    public CategoryDescriptionEmptyException(){
+        super("the category description is empty");
+    }
+}
