+++ src/main/java/bootcampragma/emazon/infrastructure/exception/CategoryNameEmptyException.java
@@ -1,0 +1,7 @@
+package bootcampragma.emazon.infrastructure.exception;
+
+public class CategoryNameEmptyException extends RuntimeException{
+    public CategoryNameEmptyException(){
+        super("The category name is empty");
+    }
+}
