+++ src/main/java/bootcampragma/emazon/infrastructure/exception/CategoryOversizeDescriptionException.java
@@ -1,0 +1,7 @@
+package bootcampragma.emazon.infrastructure.exception;
+
+public class CategoryOversizeDescriptionException extends RuntimeException{
+    public CategoryOversizeDescriptionException(){
+        super("The category description is too long");
+    }
+}
