+++ src/main/java/bootcampragma/emazon/infrastructure/exception/CategoryOversizeNameException.java
@@ -1,0 +1,7 @@
+package bootcampragma.emazon.infrastructure.exception;
+
+public class CategoryOversizeNameException extends RuntimeException{
+    public CategoryOversizeNameException(){
+        super("The category name is too long");
+    }
+}
