+++ src/main/java/bootcampragma/emazon/infrastructure/output/jpa/adapter/CategoryJpaAdapter.java
@@ -1,0 +1,38 @@
+package bootcampragma.emazon.infrastructure.output.jpa.adapter;
+
+import bootcampragma.emazon.domain.entity.Category;
+import bootcampragma.emazon.domain.spi.ICategoryPersistencePort;
+import bootcampragma.emazon.infrastructure.exception.*;
+import bootcampragma.emazon.infrastructure.output.jpa.mapper.CategoryEntityMapper;
+import bootcampragma.emazon.infrastructure.output.repository.ICategoryRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class CategoryJpaAdapter implements ICategoryPersistencePort {
+
+    private final ICategoryRepository categoryRepository;
+    private final CategoryEntityMapper categoryEntityMapper;
+
+
+    @Override
+    public void saveCategory(Category category) {
+        if (categoryRepository.existsByName(category.getName())) {
+            throw new CategoryAlreadyExistException();
+        }
+        if(category.getName().isEmpty()){
+            throw new CategoryNameEmptyException();
+        }
+        if(category.getDescription().isEmpty()){
+            throw new CategoryDescriptionEmptyException();
+        }
+        if(category.getName().length() > Category.MAX_NAME_LENGTH){
+            throw new CategoryOversizeNameException();
+        }
+        if(category.getDescription().length() > Category.MAX_DESCRIPTION_LENGTH){
+            throw new CategoryOversizeDescriptionException();
+        }
+        categoryRepository.save(categoryEntityMapper.toCategoryEntity(category));
+    }
+}
