+++ src/main/java/bootcampragma/emazon/infrastructure/output/jpa/mapper/CategoryEntityMapper.java
@@ -1,0 +1,16 @@
+package bootcampragma.emazon.infrastructure.output.jpa.mapper;
+
+import bootcampragma.emazon.domain.entity.Category;
+import bootcampragma.emazon.infrastructure.output.jpa.entity.CategoryEntity;
+import org.mapstruct.Mapper;
+import org.mapstruct.ReportingPolicy;
+
+@Mapper(componentModel = "spring",
+    unmappedTargetPolicy = ReportingPolicy.IGNORE,
+    unmappedSourcePolicy = ReportingPolicy.IGNORE)
+public interface CategoryEntityMapper  {
+
+    CategoryEntity toCategoryEntity(Category category);
+    Category toDomainCategory(CategoryEntity category);
+
+}
