+++ src/main/java/bootcampragma/emazon/infrastructure/output/repository/ICategoryRepository.java
@@ -1,0 +1,9 @@
+package bootcampragma.emazon.infrastructure.output.repository;
+
+import bootcampragma.emazon.infrastructure.output.jpa.entity.CategoryEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ICategoryRepository extends JpaRepository<CategoryEntity, Long> {
+    // Método para verificar si existe una categoría por nombre
+    boolean existsByName(String name);
+}
