+++ src/main/resources/application.properties
@@ -1,9 +1,11 @@
 spring.application.name=emazon
-server.port=9090
 
 
-spring.datasource.url=jdbc:mysql://localhost:3306/microservicestock
-spring.datasource.username=root
-spring.datasource.password=root
+server.port=${PORT}
+
+
+spring.datasource.url=${URL}
+spring.datasource.username=${USER_NAME}
+spring.datasource.password=${PASSWORD}
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.hibernate.ddl-auto=update
