--- src/test/java/BootcampPragma/emazon/EmazonApplicationTests.java
@@ -1,13 +1,0 @@
-package BootcampPragma.emazon;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EmazonApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
